Commits/PRs: Write new guidelines for commit messages.
This replaces the existing commit guidelines, to be used by all jQuery projects,
replacing the rules that jQuery UI used. This commit message itself follows
those new rules as well.

It changes usage of 'ticket' to 'issue' in one other place, to have a
consistent term instead of a mix of both.

Fixes gh-59